+++ locals.tf
@@ -8,6 +8,9 @@ locals {
     stack = var.stack
   }
 
+  administrator_login = format("%s@%s", azurerm_mysql_server.mysql_server.administrator_login, azurerm_mysql_server.mysql_server.name)
+  db_users_login      = formatlist("%s@%s", mysql_user.users.*.user, azurerm_mysql_server.mysql_server.name)
+
   tier_map = {
     "GeneralPurpose"  = "GP"
     "Basic"           = "B"