+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 Added
   * AZ-44 + TER-306: First release
+  * AZ-128: Create databases users